Changes have been made to the restaurant and menu view. Minor changes have been made to Entities. Added the ability to create orders (cart simulation).

+++ src/main/java/com/foodapp/model/user/User.java
@@ -37,5 +37,5 @@ public class User {
 
     )
     private List<Rating> ratings = new ArrayList<>();
-    private Integer extraPoints;
+    private Double extraPoints;
 }
